+++ packages/core/src/contexts/transaction.ts
@@ -1,6 +1,7 @@
 import * as db from '@nawadi/db'
 import { AsyncLocalStorage } from 'async_hooks'
-import { useDatabase } from './database.js'
+import postgres from 'postgres'
+import { useDatabase, withDatabase } from './database.js'
 
 // Initializes an instance of AsyncLocalStorage to store database transaction contexts.
 const storage = new AsyncLocalStorage<db.Transaction>()
@@ -39,4 +40,55 @@ export async function withTransaction<T>(job: () => Promise<T>): Promise<T> {
 export function useTransaction(): db.Transaction | undefined {
   const context = storage.getStore()
   return context
+}
+
+/**
+ *
+ * Migrates the local supabase database to the latest version and then runs the job
+ * with a database context that points to the local supabase database. The job is run
+ * in a transaction that is rolled back at the end of the job so that the database
+ * is kept clean
+ *
+ * @param job async test job to run
+ * @returns whatever the job returns
+ */
+export async function withTestTransaction<T>(
+  job: () => Promise<T>,
+): Promise<T> {
+  const pgUri =
+    process.env.PGURI ??
+    'postgresql://postgres:postgres@127.0.0.1:54322/postgres'
+
+  {
+    // use a single connection pool for the migration
+    const pgSql = postgres(pgUri.toString(), {
+      max: 1,
+    })
+    try {
+      const database = db.createDatabase(pgSql)
+      // migrate (set up) the database
+      await db.migrateDatabase(database)
+    } finally {
+      await pgSql.end()
+    }
+  }
+
+  const result = await withDatabase({ pgUri }, async () =>
+    withTransaction(async () => {
+      try {
+        const result = await job()
+        return result
+      } finally {
+        const transaction = useTransaction()
+        try {
+          transaction!.rollback()
+        } catch (error) {
+          // rollback will throw a Rollback error this is expected behavior and
+          // therefore should not throw
+        }
+      }
+    }),
+  )
+
+  return result
 }
